requirements.txt
acme==0.31.0
asgiref==3.2.3
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==19.2
colorama==0.3.7
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==2.1
django-hosts==3.0
django-jsonfield==1.3.1
django-markdownx==2.0.28
django-widget-tweaks==1.4.5
docopt==0.6.2
ec2-hibinit-agent==1.0.0
future==0.15.2
googletrans==2.4.0
gunicorn==20.0.4
hibagent==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10
josepy==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
Markdown==3.1.1
MarkupSafe==1.0
mistune==0.8.4
mock==2.0.0
ndg-httpsclient==0.4.4
netifaces==0.10.4
numpy==1.17.4
oauthlib==2.0.6
PAM==0.4.2
parsedatetime==2.4
pbr==3.1.1
Pillow==6.2.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.4
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
sqlparse==0.3.0
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==16.7.8
yarg==0.1.9
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
social-auth-app-django==3.1.0
djangorestframework==3.11.0
django-cors-headers==3.2.1
